+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## [Unreleased]
+### Changed
+* Changed the hes event filter to align with other new event filter
+### Added
+* Added event filters for birth, death, pregnancy loss and MSDS
 
 ## 5.9.0 / 2023-01-17
 * Added 'other' option to anomaly screening status filter (#90)